Jumping!
John:
After a very long time I have finally implemented jump physics.
This is something that the game has needed ever since the beginning.
now that the player can finally jump instead of flying I can now work on
player health, enemys and proper platforming.

+++ ArmouredWurm/src/game/Looper.java
@@ -14,7 +14,7 @@ public class Looper extends Platform
 	}
 	public void update(World theWorld)
 	{
-		x = x + speedX;
+		x = (int) (x + speedX);
 		y = theWorld.getY() + trueY;
 	}
 	public void render(Graphics g)
